pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slidebake"
version = "0.1.0"
description = "Bake slide-deck PDFs into page-by-page Markdown with OCR and optional translation."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "slidebake contributors" },
]
dependencies = [
  "openai>=2.0.0",
  "pillow>=11.0.0",
  "pymupdf>=1.24.0",
  "rapidocr-onnxruntime>=1.4.0",
  "rich>=13.0.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.8.0",
]

[project.scripts]
slidebake = "slidebake.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/slidebake"]

[tool.hatch.build.targets.sdist]
include = [
  "/src/slidebake",
  "/tests",
  "/README.md",
  "/pyproject.toml",
  "/.gitignore",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
